requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-core>=0.1.0
pyyaml>=6.0
markdownify>=0.11.0
pandas>=2.0.0
PyGithub>=2.0.0
tabulate>=0.9.0
python-dotenv>=1.0.0
openai>=1.0.0